Return error instead of panicking
Although these panics should never occur, they make downstream consumers
nervous and have led to users forking the code. So the code now returns an
error which should never be non-nil and it is up to the caller to handle it
appropriately.

Also fix a few lint and formatting issues.